Add files via upload
rename reports to start with v1 for go-live. rc=release candidate. this can be dropped once we're sure no changes need to be made prior to July 1st. 
Variant report v1.0.0-alpha.1 will be deprecated with v1.0.0-rc.1 once rc.1 is validated. 
Variant report v1.0.0-rc.1 combines the cDNA and protein fields into a single field since we defined in the VCF. This allows us to store the legacy nomenclature in the VCF and pull it in to the report.